TEMP fix development settings until official fix is found.
Issue is with the restricted settings.. specifically oem unlock and
keep screen on while charging.. These features seem rather pointless..
obviously were oem unlocked and it needs to stay that way and why in
the world would you keep the screen on while charging?  So.. simple fix..
remove the options causing the issue.  May revert this once an official
fix is found, but in the mean time, dev settings are working again.

Change-Id: Ic8d7deb4848f7fa0362e01e0917aa454a57ab2f7
Signed-off-by: scotthartbti <scott.hart.bti@gmail.com>